% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_returns.R
\name{accumulate_returns}
\alias{accumulate_returns}
\title{accumulate_returns}
\usage{
accumulate_returns(returns, roll = 1, periods_on = "end")
}
\arguments{
\item{returns}{xts containing returns}

\item{roll}{default is 1}

\item{periods_on}{default is NULL. It can take any
value that is supported by xts::endpoints}
}
\description{
Accumulates returns
}
\details{
xx
}

