% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_returns.R
\name{aggregate_returns}
\alias{aggregate_returns}
\title{aggregate_returns}
\usage{
aggregate_returns(returns)
}
\arguments{
\item{returns}{xts containing returns}
}
\description{
Aggregates returns
}
\details{
xx
}

