% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_returns.R
\name{annualize_returns}
\alias{annualize_returns}
\title{annualize_returns}
\usage{
annualize_returns(returns, factor = 252)
}
\arguments{
\item{returns}{xts containing returns}

\item{factor}{annualization factor}
}
\description{
Annualize returns
}
\details{
xx
}

