% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_assetId.R
\name{get_assetId}
\alias{get_assetId}
\title{get_assetId}
\usage{
get_assetId(tipo, emisora, serie)
}
\arguments{
\item{tipo}{character vector containing 'tipo valor'.}

\item{emisora}{character vector containing 'emisora'.}

\item{serie}{character vector conatining 'serie'.}
}
\value{
xx
}
\description{
Creates a unique id from 'tipo valor',
'emisora' and 'serie'.
}
\details{
xx
}
\examples{
get_AssetId("M", "BONOS", "191211")
}

