% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimist.R
\name{get_contribParity_weights}
\alias{get_contribParity_weights}
\title{get_contribParity_weights}
\usage{
get_contribParity_weights(dlyChg, halflife)
}
\arguments{
\item{dlyChg}{xts that contains daily changes}

\item{halflife}{number of days that accumulate 0.5 density}
}
\description{
gets risk contribution parity weights
}
