% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimist.R
\name{get_covRet_from_dlyChg}
\alias{get_covRet_from_dlyChg}
\title{get_covRet_from_dlyChg}
\usage{
get_covRet_from_dlyChg(dlyChg, halflife = NULL, ...)
}
\arguments{
\item{dlyChg}{xts that contains daily changes}

\item{halflife}{number of days that accumulate 0.5 density}

\item{...}{parameters passed to get_w_with_geomTruncDecay()}
}
\description{
builds weighted covariance matrix
}
