% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimist.R
\name{get_cumRet_from_dlyChg}
\alias{get_cumRet_from_dlyChg}
\title{get_cumRet_from_dlyChg}
\usage{
get_cumRet_from_dlyChg(dlyChg)
}
\arguments{
\item{dlyChg}{xts containing daily changes}
}
\description{
Builds an xts containing cumulative returns
}
