% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimist.R
\name{get_dlyChg_from_price}
\alias{get_dlyChg_from_price}
\title{get_dlyChg_from_price}
\usage{
get_dlyChg_from_price(price)
}
\arguments{
\item{price}{xts containing prices}
}
\description{
Builds an xts containing daily price changes
}
