% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimist.R
\name{get_dlyImpliedRet_from_dlyWeights}
\alias{get_dlyImpliedRet_from_dlyWeights}
\title{get_dlyImpliedRet_from_dlyWeights}
\usage{
get_dlyImpliedRet_from_dlyWeights(dlyWeights, dlyChg, halflife,
  min_window = ceiling(halflife * 2.3), roll_parity = NULL)
}
\arguments{
\item{dlyWeights}{xts that contains portfolio daily weights}

\item{dlyChg}{xts that contains daily changes}

\item{halflife}{number of days that accumulate 0.5 density}

\item{min_window}{number of days in the first window}

\item{roll_parity}{days for the output's the rolling window}
}
\description{
gets implied returns from portfolio weights
}
