% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimist.R
\name{get_meanRet_from_dlyChg}
\alias{get_meanRet_from_dlyChg}
\title{get_meanRet_from_dlyChg}
\usage{
get_meanRet_from_dlyChg(dlyChg, method = "geometric", halflife = NULL,
  interval = c(1e-06, 0.999999), annualization_factor = 1)
}
\arguments{
\item{dlyChg}{xts that contains daily changes}

\item{method}{either 'arithmetic' or 'geometric'}

\item{halflife}{number of days that accumulate 0.5 density}

\item{interval}{as in uniroot}

\item{annualization_factor}{in days (252 days in a year)}
}
\description{
Builds weights with geometric truncated decay
}
