% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimist.R
\name{get_optim_dlyWeights_from_dlyImpliedRet}
\alias{get_optim_dlyWeights_from_dlyImpliedRet}
\title{get_optim_dlyWeights_from_dlyImpliedRet}
\usage{
get_optim_dlyWeights_from_dlyImpliedRet(dlyImpliedRet, dlyChg, halflife,
  min_window = ceiling(halflife * 2.3), roll_optim = NULL, delta = 1)
}
\arguments{
\item{dlyImpliedRet}{xts that contains portfolio daily implied returns}

\item{dlyChg}{xts that contains daily changes}

\item{halflife}{number of days that accumulate 0.5 density}

\item{min_window}{number of days in the first window}

\item{roll_optim}{days for the output's the rolling window}

\item{delta}{risk aversion parameter}
}
\description{
gets daily weights for optimal portolio from daily implied returns
}
