% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_prices.R
\name{get_prices_csv}
\alias{get_prices_csv}
\title{get_prices_csv}
\usage{
get_prices_csv(file)
}
\arguments{
\item{file}{name of the file from which the data
will be read from}
}
\description{
Builds an xts containing asset prices
from a csv.
}
\details{
Date should be provided on the first column
with yyyy-mm-dd format
}

