% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimist.R
\name{get_prices_yahoo}
\alias{get_prices_yahoo}
\title{get_prices_yahoo}
\usage{
get_prices_yahoo(yahoo_id, column = "Adjusted", from = "2007-01-01",
  to = Sys.Date(), ..., periodicity = "daily")
}
\arguments{
\item{yahoo_id}{character array containining yahoo ids.}

\item{column}{indicating which column should be returned: "Open", "High", "Low", "Close", "Volume", ("Adjusted")}

\item{from}{start date. Retrieve data no earlier than this date. (2007-01-01).}

\item{to}{end date. (Sys.Date())}

\item{...}{additional parameters}

\item{periodicity}{periodicity of data to query and return. Must be one of "daily", "weekly", "monthly", ("daily")}
}
\description{
Builds an xts containing prices downloaded from Yahoo! Finance.
}
