% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimist.R
\name{get_principalParity_weights}
\alias{get_principalParity_weights}
\title{get_principalParity_weights}
\usage{
get_principalParity_weights(dlyChg, halflife, num_factors = 3)
}
\arguments{
\item{dlyChg}{xts that contains daily changes}

\item{halflife}{number of days that accumulate 0.5 density}
}
\description{
gets risk contribution parity weights
}
