% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimist.R
\name{get_rebalance}
\alias{get_rebalance}
\title{get_rebalance}
\usage{
get_rebalance(dlyChg, rebWeight)
}
\arguments{
\item{dlyChg}{xts that contains daily changes}

\item{rebWeight}{xts that contains rebalance weights}
}
\description{
Computes daily portofolio weights, value and contributions
}
