% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimist.R
\name{get_rebalance_}
\alias{get_rebalance_}
\title{get_rebalance_}
\usage{
get_rebalance_(dlyChg, rebWeight)
}
\arguments{
\item{dlyChg}{xts that contains daily changes}

\item{rebWeight}{xts that contains rebalance weights}
}
\description{
Computes daily portofolio weights and value
}
