% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimist.R
\name{get_rebalance_t2}
\alias{get_rebalance_t2}
\title{get_rebalance_t2}
\usage{
get_rebalance_t2(dlyChg_t2, portWeightOpen_t2, portValue_t1)
}
\arguments{
\item{dlyChg_t2}{daily changes at t2}

\item{portWeightOpen_t2}{portfolio weight at the open on t2}

\item{portValue_t1}{portfolio value at t1}
}
\description{
Gets portfolio weights and value at time t2
}
