% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_returns.R
\name{get_returns}
\alias{get_returns}
\title{get_returns}
\usage{
get_returns(prices, roll = 1, periods_on = "days")
}
\arguments{
\item{roll}{default is 1}

\item{periods_on}{default is 'days'}
}
\description{
Creates xts of returns
}
\details{
xx
}

