% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimist.R
\name{get_riskParity_dlyWeights}
\alias{get_riskParity_dlyWeights}
\title{get_riskParity_dlyWeights}
\usage{
get_riskParity_dlyWeights(dlyChg,
  fun_riskParity_weights = get_naiveParity_weights, halflife,
  min_window = ceiling(halflife * 2.3), roll_parity = NULL, ...)
}
\arguments{
\item{dlyChg}{xts that contains daily changes}

\item{fun_riskParity_weights}{function to compute risk parity weights}

\item{halflife}{number of days that accumulate 0.5 density}

\item{min_window}{number of days in the first window}

\item{roll_parity}{days for the output's the rolling window}

\item{...}{fun_riskParity_weights() parameters}
}
\description{
Builds an `xts` that contains daily risk parity weights
}
