% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimist.R
\name{get_rollChg_from_dlyChg}
\alias{get_rollChg_from_dlyChg}
\title{get_rollChg_from_dlyChg}
\usage{
get_rollChg_from_dlyChg(dlyChg, method = "geometric", roll = 30,
  halflife = NULL, trim = FALSE, interval = c(1e-06, 0.999999))
}
\arguments{
\item{dlyChg}{xts that contains daily changes}

\item{method}{either 'arithmetic' or 'geometric'}

\item{roll}{number of days in a rolling window}

\item{halflife}{number of days that accumulate 0.5 density}

\item{trim}{should the first rolling window be eliminated?}

\item{interval}{as in uniroot}
}
\description{
computes rolling returns
}
