% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_VectorAnalitico.R
\name{get_VectorAnalitico}
\alias{get_VectorAnalitico}
\title{get_VectorAnalitico}
\usage{
get_VectorAnalitico(file_VectorAnalitico, file_Monedas = NULL)
}
\arguments{
\item{file_VectorAnalitico}{file name of csv containing
VectorAnalitico.}

\item{file_Monedas}{file name of csv containing Monedas.}
}
\value{
xx
}
\description{
Creates a data.frame from a csv file
containing VectorAnalitico.
}
\details{
Assume file_VectorAnalitico for a single date.
}
\examples{
get_VectorAnalitico("VectorAnalitico20161130MD.csv")
}

