% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimist.R
\name{get_w_with_geomTruncDecay}
\alias{get_w_with_geomTruncDecay}
\title{get_w_with_geomTruncDecay}
\usage{
get_w_with_geomTruncDecay(T, halflife, interval = c(1e-06, 0.999999))
}
\arguments{
\item{T}{where the geometric distribution is truncated}

\item{halflife}{number of days that accumulate 0.5 density}

\item{interval}{as in uniroot}
}
\description{
Builds weights with geometric truncated decay
}
