% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_returns.R
\name{mean_returns}
\alias{mean_returns}
\title{mean_returns}
\usage{
mean_returns(returns, roll = 12, periods_on = "days", halflife = 1)
}
\arguments{
\item{returns}{xts containing mean returns}

\item{roll}{default is 12}

\item{periods_on}{default is "days"}

\item{halflife}{default is 1. As percentage}
}
\description{
Compute mean returns
}
\details{
xx
}

