% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimist.R
\name{plot_riskReward_from_dlyChg}
\alias{plot_riskReward_from_dlyChg}
\title{plot_riskReward_from_dlyChg}
\usage{
plot_riskReward_from_dlyChg(dlyChg, roll, roll_halflife = NULL)
}
\arguments{
\item{dlyChg}{xts that contains daily changes}

\item{roll}{number of days in a rolling window}

\item{roll_halflife}{within a rolling window}
}
\description{
plots risk-reward
}
