% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimist.R
\name{weighted.sd}
\alias{weighted.sd}
\title{weighted.sd}
\usage{
weighted.sd(x, w, na.rm = TRUE)
}
\arguments{
\item{x}{a vector or dataframe containing the values
whose weighted standard deviation is to be computed.}

\item{w}{a vector of weights.}

\item{na.rm}{remove NA values before processing,
the default value is FALSE.}
}
\description{
Compute a weighted standard deviation.
}
\details{
See radiant.data::weighted.sd and
stats::weighted.mean.default.
}
